Sends reactions to own message
Right now it can save reactions fine and add them to YOUR message when you run a command, but what I'm still trying to figure out is how to add reactions to someone else's message. One super easy way would be to have this take commands in the CLI but that would defeat the purpose of the ease of use right in Discord.

I need to make the /list command pretty, add support for Nitro users with their custom emojis, and most importantly fix the above problem. Also going to add a pretty README as soon as I do all of that.