Renamed main file
Because of the library React, it wasn't detecting my project as having any languages, in it, so I didn't the pretty yellow underline at the top of the README. That was unacceptable, so here we are. Also added some entirely untested code that probably works to prevent any funky issues from popping up down the line.