Avoided deleted message problem
I was trying to be fancy and get around deleted messages by reacting to the last message the user sent that is still not deleted, but no matter what I do with promises there's always one emoji in the chain that fails to react. I don't know what the hell the problem is and I'm sick of tearing my hair out over it, so I commented out my working code in case someone else wants to take a shot at solving it.