Updated README.md
Removed some TODOs that I accomplished and stuff that I no longer feel is within the scope of this project.

Also, I've had enough people message me about not being able to run npm that I figured it was best to make it clear that you have to have Node installed to use Node. I figured that much would be self-explanatory, but that's on me for overestimating the user. Always a mistake.